requirements.txt
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.7.4
fastapi==0.111.0
langchain==0.3.0
langchain-openai==0.2.0